MAINT: swap int for long type for bdt{r,ri,rc} functions
  * in PR 11045 we kept the types int. Josh left a comment
    before closing gh-issue 9454 about switching to longs
    which was a suggestion to improve numerics of bdtr in
    some cases deep in the tail of binom cdf. This PR
    makes the int types into long type for these 3 functions.

+++ scipy/special/_legacy.pxd
@@ -62,21 +62,21 @@ cdef inline double bdtr_unsafe(double k, double n, double p) nogil:
     if npy_isnan(n) or npy_isinf(n):
         return nan
     else:
-        return bdtr(k, <int>n, p)
+        return bdtr(k, <long>n, p)
 
 cdef inline double bdtrc_unsafe(double k, double n, double p) nogil:
     _legacy_deprecation("bdtrc", k, n)
     if npy_isnan(n) or npy_isinf(n):
         return nan
     else:
-        return bdtrc(k, <int>n, p)
+        return bdtrc(k, <long>n, p)
 
 cdef inline double bdtri_unsafe(double k, double n, double p) nogil:
     _legacy_deprecation("bdtri", k, n)
     if npy_isnan(n) or npy_isinf(n):
         return nan
     else:
-        return bdtri(k, <int>n, p)
+        return bdtri(k, <long>n, p)
 
 cdef inline double expn_unsafe(double n, double x) nogil:
     if npy_isnan(n):
